feat(MonitoringServiceCLI): add meter reading submission logic and validation
- Added logic for submitting meter readings in MonitoringServiceCLI
- Implemented method in Repository for adding meter readings to the database
- Created MeterReadingIndicationValidation for validating meter readings
- Introduced MeterReadingExistsException for handling attempts to submit duplicate readings in the same month
- Integrated MapStruct with mapper interfaces for data transformation
- Made minor changes and improvements throughout the project

+++ src/main/java/monitoring/service/dev/models/MeterReading.java
@@ -13,7 +13,7 @@ public class MeterReading {
 
     private int id;
 
-    private int indication;
+    private double indication;
 
     private LocalDateTime date;
 }
